Wiwifixes to make the data and resource pack work after bad refactoring

+++ wiwitool/include/Minecraft_pack.hpp
@@ -6,7 +6,7 @@
 /// A Minecraft pack generates both a data- and resouce-pack.
 class Minecraft_pack {
 private:
-  struct mcversion_range { float min, max; };
+  struct mcversion_range { double min, max; };
 
   const std::string description;
   const mcversion_range data_ver{88.0, 94.1}, res_ver{73.0, 76.0};
